Writing flow: consider events only from DOM descendents
