+++ packages/block-editor/src/components/block-list/block-selection-button.js
@@ -9,17 +9,30 @@ import classnames from 'classnames';
 import { Button } from '@wordpress/components';
 import { useSelect, useDispatch } from '@wordpress/data';
 import { useEffect, useRef } from '@wordpress/element';
-import { BACKSPACE, DELETE } from '@wordpress/keycodes';
+import {
+	BACKSPACE,
+	DELETE,
+	UP,
+	DOWN,
+	LEFT,
+	RIGHT,
+	TAB,
+	ESCAPE,
+	ENTER,
+	SPACE,
+} from '@wordpress/keycodes';
 import {
 	getBlockType,
 	__experimentalGetAccessibleBlockLabel as getAccessibleBlockLabel,
 } from '@wordpress/blocks';
 import { speak } from '@wordpress/a11y';
+import { focus } from '@wordpress/dom';
 
 /**
  * Internal dependencies
  */
 import BlockTitle from '../block-title';
+import { getBlockDOMNode } from '../../utils/dom';
 
 /**
  * Returns true if the user is using windows.
@@ -28,6 +41,40 @@ import BlockTitle from '../block-title';
  */
 function isWindows() {
 	return window.navigator.platform.indexOf( 'Win' ) > -1;
+}
+
+function selector( select ) {
+	const {
+		getSelectedBlockClientId,
+		getMultiSelectedBlocksEndClientId,
+		getPreviousBlockClientId,
+		getNextBlockClientId,
+		hasBlockMovingClientId,
+		getBlockIndex,
+		getBlockRootClientId,
+		getClientIdsOfDescendants,
+		canInsertBlockType,
+		getBlockName,
+	} = select( 'core/block-editor' );
+
+	const selectedBlockClientId = getSelectedBlockClientId();
+	const selectionEndClientId = getMultiSelectedBlocksEndClientId();
+
+	return {
+		selectedBlockClientId,
+		selectionBeforeEndClientId: getPreviousBlockClientId(
+			selectionEndClientId || selectedBlockClientId
+		),
+		selectionAfterEndClientId: getNextBlockClientId(
+			selectionEndClientId || selectedBlockClientId
+		),
+		hasBlockMovingClientId,
+		getBlockIndex,
+		getBlockRootClientId,
+		getClientIdsOfDescendants,
+		canInsertBlockType,
+		getBlockName,
+	};
 }
 
 /**
@@ -82,12 +129,115 @@ function BlockSelectionButton( { clientId, rootClientId, ...props } ) {
 		}
 	}, [] );
 
+	const {
+		selectedBlockClientId,
+		selectionBeforeEndClientId,
+		selectionAfterEndClientId,
+		hasBlockMovingClientId,
+		getBlockIndex,
+		getBlockRootClientId,
+		getClientIdsOfDescendants,
+	} = useSelect( selector, [] );
+	const {
+		selectBlock,
+		clearSelectedBlock,
+		setBlockMovingClientId,
+		moveBlockToPosition,
+	} = useDispatch( 'core/block-editor' );
+
 	function onKeyDown( event ) {
 		const { keyCode } = event;
+		const isUp = keyCode === UP;
+		const isDown = keyCode === DOWN;
+		const isLeft = keyCode === LEFT;
+		const isRight = keyCode === RIGHT;
+		const isTab = keyCode === TAB;
+		const isEscape = keyCode === ESCAPE;
+		const isEnter = keyCode === ENTER;
+		const isSpace = keyCode === SPACE;
+		const isShift = event.shiftKey;
 
 		if ( keyCode === BACKSPACE || keyCode === DELETE ) {
 			removeBlock( clientId );
 			event.preventDefault();
+			return;
+		}
+
+		const navigateUp = ( isTab && isShift ) || isUp;
+		const navigateDown = ( isTab && ! isShift ) || isDown;
+		// Move out of current nesting level (no effect if at root level).
+		const navigateOut = isLeft;
+		// Move into next nesting level (no effect if the current block has no innerBlocks).
+		const navigateIn = isRight;
+
+		let focusedBlockUid;
+		if ( navigateUp ) {
+			focusedBlockUid = selectionBeforeEndClientId;
+		} else if ( navigateDown ) {
+			focusedBlockUid = selectionAfterEndClientId;
+		} else if ( navigateOut ) {
+			focusedBlockUid =
+				getBlockRootClientId( selectedBlockClientId ) ??
+				selectedBlockClientId;
+		} else if ( navigateIn ) {
+			focusedBlockUid =
+				getClientIdsOfDescendants( [ selectedBlockClientId ] )[ 0 ] ??
+				selectedBlockClientId;
+		}
+		const startingBlockClientId = hasBlockMovingClientId();
+
+		if ( isEscape && startingBlockClientId ) {
+			setBlockMovingClientId( null );
+		}
+		if ( ( isEnter || isSpace ) && startingBlockClientId ) {
+			const sourceRoot = getBlockRootClientId( startingBlockClientId );
+			const destRoot = getBlockRootClientId( selectedBlockClientId );
+			const sourceBlockIndex = getBlockIndex(
+				startingBlockClientId,
+				sourceRoot
+			);
+			let destinationBlockIndex = getBlockIndex(
+				selectedBlockClientId,
+				destRoot
+			);
+			if (
+				sourceBlockIndex < destinationBlockIndex &&
+				sourceRoot === destRoot
+			) {
+				destinationBlockIndex -= 1;
+			}
+			moveBlockToPosition(
+				startingBlockClientId,
+				sourceRoot,
+				destRoot,
+				destinationBlockIndex
+			);
+			selectBlock( startingBlockClientId );
+			setBlockMovingClientId( null );
+		}
+		if ( navigateDown || navigateUp || navigateOut || navigateIn ) {
+			if ( focusedBlockUid ) {
+				event.preventDefault();
+				selectBlock( focusedBlockUid );
+			} else if ( isTab && selectedBlockClientId ) {
+				const wrapper = getBlockDOMNode(
+					selectedBlockClientId,
+					document
+				);
+				let nextTabbable;
+
+				if ( navigateDown ) {
+					nextTabbable = focus.tabbable.findNext( wrapper );
+				} else {
+					nextTabbable = focus.tabbable.findPrevious( wrapper );
+				}
+
+				if ( nextTabbable ) {
+					event.preventDefault();
+					nextTabbable.focus();
+					clearSelectedBlock();
+				}
+			}
 		}
 	}
