+++ packages/block-editor/src/components/block-list/block-wrapper.js
@@ -71,13 +71,21 @@ export function useBlockProps( props = {}, { __unstableIsHtml } = {} ) {
 		initialPosition,
 		shouldFocusFirstElement,
 		isNavigationMode,
+		isBlockMovingMode,
+		canInsertMovingBlock,
 	} = useSelect(
 		( select ) => {
 			const {
 				getSelectedBlocksInitialCaretPosition,
 				isMultiSelecting: _isMultiSelecting,
 				isNavigationMode: _isNavigationMode,
+				hasBlockMovingClientId,
+				canInsertBlockType,
+				getBlockName,
+				getBlockRootClientId,
 			} = select( 'core/block-editor' );
+
+			const movingClientId = hasBlockMovingClientId();
 
 			return {
 				shouldFocusFirstElement:
@@ -88,9 +96,17 @@ export function useBlockProps( props = {}, { __unstableIsHtml } = {} ) {
 					? getSelectedBlocksInitialCaretPosition()
 					: undefined,
 				isNavigationMode: _isNavigationMode,
+				isBlockMovingMode: isSelected && !! movingClientId,
+				canInsertMovingBlock:
+					isSelected &&
+					movingClientId &&
+					canInsertBlockType(
+						getBlockName( movingClientId ),
+						getBlockRootClientId( clientId )
+					),
 			};
 		},
-		[ isSelected ]
+		[ isSelected, clientId ]
 	);
 	const { insertDefaultBlock, removeBlock, selectBlock } = useDispatch(
 		'core/block-editor'
@@ -331,7 +347,11 @@ export function useBlockProps( props = {}, { __unstableIsHtml } = {} ) {
 			className,
 			props.className,
 			wrapperProps.className,
-			{ 'is-hovered': isHovered }
+			{
+				'is-hovered': isHovered,
+				'is-block-moving-mode': isBlockMovingMode,
+				'can-insert-moving-block': canInsertMovingBlock,
+			}
 		),
 		style: { ...wrapperProps.style, ...props.style },
 	};
